API/requirements.txt
annoy~=1.17.0
matplotlib~=3.5.0
pandas~=1.3.4
numpy~=1.21.4
tensorflow~=2.7.0
tqdm~=4.62.3
Pillow~=8.4.0
opencv-python
scikit-image

fastapi~=0.70.0
uvicorn[standard]~=0.15.0
aiosqlite
sqlalchemy~=1.4.3
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart
keras
dash
pandas
starlette~=0.16.0
json2html~=1.3.0
pydantic~=1.8.2